convert spotirec to module
